+++ src/pages/home/sections/section1/Section1.css
@@ -165,18 +165,7 @@
   margin-right: auto;
   margin-bottom: 12px;
 }
-.codeutsava__section1-intro-audio-div{
-  margin-top: 10px;
-}
-button.codeutsava__section1-intro-audio-btn {
-  background-color: rgb(244, 117, 21);
-  border-radius: 4px;
-  padding: 2px;
-  padding-top:4px ;
-  border: none;
-  outline: none;
-  cursor: pointer
-}
+
 .codeutsava__section1-side-links li button:hover {
   list-style: none;
   outline: none;
